Add: Server recibe tambien la agencia del cliente. Fix: Se modifica el protocolo para mandar el len como uint8

+++ client/common/sender.go
@@ -6,24 +6,26 @@ import (
 )
 
 // SendSerializedBet sends:
-// - len(nombre) in 4 bytes
+// - len(nombre) in 1 byte
 // - nombre
-// - len(apellido) in 4 bytes
+// - len(apellido) in 1 byte
 // - apellido
-// - len(dni) in 4 bytes
+// - len(dni) in 1 byte
 // - dni
-// - len(nacimiento) in 4 bytes
+// - len(nacimiento) in 1 byte
 // - nacimiento
-// - len(numero) in 4 bytes
+// - len(numero) in 1 byte
 // - numero
+// - len(agencia) in 1 byte
+// - agencia
 func SendSerializedBet(conn net.Conn, bet SerializedBet) error {
 	for _, data := range bet.data {
-		err := SendLen(conn, uint32(len(data)))
+		err := SendLen(conn, uint8(len(data)))
 		if err != nil {
 			return err
 		}
 
-		err = SendData(conn, data, uint32(len(data)))
+		err = SendData(conn, data, uint8(len(data)))
 		if err != nil {
 			return err
 		}
@@ -32,7 +34,7 @@ func SendSerializedBet(conn net.Conn, bet SerializedBet) error {
 }
 
 // SendLen sends a uint32 through the connection
-func SendLen(conn net.Conn, data uint32) error {
+func SendLen(conn net.Conn, data uint8) error {
 	err := binary.Write(conn, binary.BigEndian, data)
 	if err != nil {
 		log.Criticalf("action: SendLen | result: fail | error: %v", err)
@@ -43,15 +45,15 @@ func SendLen(conn net.Conn, data uint32) error {
 }
 
 // SendData sends bytesToSend bytes of data through the connection
-func SendData(conn net.Conn, data []byte, bytesToSend uint32) error {
-	var total uint32 = 0
+func SendData(conn net.Conn, data []byte, bytesToSend uint8) error {
+	var total uint8 = 0
 	for total < bytesToSend {
 		n, err := conn.Write(data[total:])
 		if err != nil {
 			log.Criticalf("action: SendData | result: fail | error: Only %v of %v bytes were sent", total, bytesToSend)
 			return err
 		}
-		total += uint32(n)
+		total += uint8(n)
 	}
 	log.Infof("action: SendData | result: success | Bytes sent: %v | Data: %v", total, data)
 	return nil
